+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.0.6
+
+- Fixed the shotgun having increased damage for clients.
+- Fixed the death sound of Baboon Hawks, Hoarder Bugs & Nutcrackers not working.
+- Fixed the old bird missiles getting stuck if the owning old bird is destroyed by the Earth Leviathan.
+
 ### 1.0.5
 
 - Fixed exception if the map has a null outside object.